requirements.txt
-e ./SI_Toolkit
casadi
derivative
do-mpc==4.1.1
engineering_notation
git+https://github.com/homangab/gradcem@master#egg=mpc
GitPython
gpflow
gymnasium[classic_control,mujoco,box2d]>=0.26.0
imageio
matplotlib
natsort
nni
numba
numpy<1.23
pandas
pip
pygame
pyyaml
scienceplots
scipy
setuptools
scikit-spatial
snakeviz
sympy
tensorflow==2.10.1
tensorflow-probability==0.18.0
torch
torchvision
tqdm
typing-extensions
guildai
ruamel.yaml